+++ HISTORY.rst
@@ -5,10 +5,11 @@ Version History
 ~~~~~~~~
 - Make ``sprockets.mixins.correlation.HandlerMixin.prepare`` async
 - Add ``sprockets.mixins.correlation.CorrelationAdapter``
+- Add ``sprockets.mixins.correlation.LoggingMixin``
 
 `1.0.1`_ (31-Mar-2015)
 ~~~~~~~~~~~~~~~~~~~~~~
- - Adds ``sprockets.mixins.correlation.HandlerMixin``
+ - Add ``sprockets.mixins.correlation.HandlerMixin``
 
 
 .. _`1.1.0`: https://github.com/sprockets/sprockets.mixins.correlation/compare/1.0.1...1.1.0